Create a hash package and move fnv hashing
